Fixed + added PID 0 check
