--- find_process/find_process_linux/find_process.go
@@ -1,12 +1,0 @@
-package find_process_linux
-
-/* import (
-	// Standard Library
-	"log"
-	"os/exec"
-	"strconv"
-	"strings"
-
-	// Nyrna Packages
-	nyrna "github.com/Merrit/nyrna"
-) */
